+++ src/CofreeBot.hs
@@ -2,10 +2,12 @@ module CofreeBot
   ( module Behaviors
   , module Bot
   , module Context
+  , module Server
   , module Utils
   ) where
 
 import           CofreeBot.Bot                 as Bot
 import           CofreeBot.Bot.Behaviors       as Behaviors
 import           CofreeBot.Bot.Context         as Context
+import           CofreeBot.Server              as Server
 import           CofreeBot.Utils               as Utils
