+++ src/CofreeBot/Bot.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE ViewPatterns #-}
 {-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
 {-# HLINT ignore "Use <$>" #-}
 {-# LANGUAGE RankNTypes #-}
@@ -7,28 +6,12 @@ module CofreeBot.Bot where
 import           CofreeBot.Utils
 import qualified Control.Arrow                 as Arrow
 import qualified Control.Category              as Cat
-import           Control.Exception              ( catch
-                                                , throwIO
-                                                )
-import           Control.Lens (ifolded, view, (^.), _Just)
-import           Control.Monad
-import           Control.Monad.Except
-import           Data.Foldable
+import           Control.Monad.State (runStateT, StateT (StateT))
+import           Data.Bifunctor (Bifunctor(..))
+import           Data.Functor ((<&>))
 import           Data.Kind
-import qualified Data.Map.Strict               as Map
 import           Data.Profunctor
-import qualified Data.Text                     as T
-import qualified Data.Text.IO                  as T
-import           Network.Matrix.Client
-import           Network.Matrix.Client.Lens
-import           System.Directory               ( createDirectoryIfMissing )
-import           System.IO
-import           System.IO.Error                ( isDoesNotExistError )
-import           System.Random
-import Data.Bifunctor (Bifunctor(..))
-import Control.Monad.State (runStateT, StateT (StateT))
-import Data.Profunctor.Traversing
-import Data.Functor ((<&>))
+import           Data.Profunctor.Traversing
 
 --------------------------------------------------------------------------------
 -- Kinds
@@ -37,10 +20,8 @@ import Data.Functor ((<&>))
 type KBot = (Type -> Type) -> Type -> Type -> Type -> Type
 
 --------------------------------------------------------------------------------
--- Types
+-- Bot Action
 --------------------------------------------------------------------------------
-
-newtype Fix f = Fix { runFix :: f (Fix f) }
 
 data BotAction s o = BotAction
   { responses :: o
@@ -48,10 +29,49 @@ data BotAction s o = BotAction
   }
   deriving Functor
 
+instance (Semigroup s, Semigroup o) => Semigroup (BotAction s o) where
+  BotAction o s <> BotAction o' s' =
+    BotAction { responses = o <> o', nextState = s <> s' }
+
+instance (Monoid s, Monoid o) => Monoid (BotAction s o) where
+  mempty = BotAction { responses = mempty, nextState = mempty }
+
+instance Bifunctor BotAction where
+  bimap f g (BotAction a b) = BotAction (g a) (f b)
+
+--------------------------------------------------------------------------------
+-- Bot
+--------------------------------------------------------------------------------
+
 -- | A 'Bot' maps from some input type 'i' and a state 's' to an
 -- output type 'o' and a state 's'
 type Bot :: KBot
 newtype Bot m s i o = Bot { runBot :: i -> s -> m (BotAction s o) }
+
+instance Monad m => Cat.Category (Bot m s) where
+  id = Bot $ \a s -> pure $ BotAction a s
+
+  Bot f . Bot g = Bot $ \a s -> do
+    BotAction b s' <- g a s
+    f b s'
+
+instance Monad m => Arrow.Arrow (Bot m s) where
+  arr f = rmap f Cat.id
+  first = first'
+
+instance Functor f => Profunctor (Bot f s) where
+  dimap f g (Bot bot) = Bot $ \a -> fmap (fmap g) . bot (f a)
+
+instance Functor f => Strong (Bot f s) where
+  first' (Bot bot) = Bot $ \(a, c) -> fmap (fmap (, c)) . bot a
+
+instance Applicative f => Choice (Bot f s) where
+  left' (Bot bot) = Bot $ either ((fmap . fmap . fmap) Left . bot)
+                                 (\c s -> pure $ BotAction (Right c) s)
+
+--------------------------------------------------------------------------------
+-- Behavior
+--------------------------------------------------------------------------------
 
 newtype Behavior m i o = Behavior { runBehavior :: i -> m (BotAction (Behavior m i o) o) }
 
@@ -80,55 +100,14 @@ instance Monad m => Traversing (Behavior m)
     $ flip runStateT b
     $ traverse (\i -> StateT $ \(Behavior b') -> fmap (\case { BotAction {..} -> (responses, nextState) }) $ b' i) is
 
-fixBot :: Functor m => Bot m s i o -> s -> Behavior m i o
-fixBot (Bot b) = go
-  where
-  go s = Behavior $ \i -> first go <$> b i s
-
-type Env :: KBot
-newtype Env m s o i = Env { runEnv :: s -> (i, o -> m s) }
-
-newtype Server m o i = Server { runServer :: (i, o -> m (Server m o i)) }
-
-fixEnv :: Functor m => Env m s o i -> s -> Server m o i
-fixEnv (Env b) = go
-  where
-  go s = Server $ fmap (fmap (fmap go)) $ b s
-
 --------------------------------------------------------------------------------
--- Instances
+-- Fix
 --------------------------------------------------------------------------------
 
-instance (Semigroup s, Semigroup o) => Semigroup (BotAction s o) where
-  BotAction o s <> BotAction o' s' =
-    BotAction { responses = o <> o', nextState = s <> s' }
+newtype Fix f = Fix { runFix :: f (Fix f) }
 
-instance (Monoid s, Monoid o) => Monoid (BotAction s o) where
-  mempty = BotAction { responses = mempty, nextState = mempty }
-
-instance Bifunctor BotAction where
-  bimap f g (BotAction a b) = BotAction (g a) (f b)
-
-instance Monad m => Cat.Category (Bot m s) where
-  id = Bot $ \a s -> pure $ BotAction a s
-
-  Bot f . Bot g = Bot $ \a s -> do
-    BotAction b s' <- g a s
-    f b s'
-
-instance Monad m => Arrow.Arrow (Bot m s) where
-  arr f = rmap f Cat.id
-  first = first'
-
-instance Functor f => Profunctor (Bot f s) where
-  dimap f g (Bot bot) = Bot $ \a -> fmap (fmap g) . bot (f a)
-
-instance Functor f => Strong (Bot f s) where
-  first' (Bot bot) = Bot $ \(a, c) -> fmap (fmap (, c)) . bot a
-
-instance Applicative f => Choice (Bot f s) where
-  left' (Bot bot) = Bot $ either ((fmap . fmap . fmap) Left . bot)
-                                 (\c s -> pure $ BotAction (Right c) s)
+loop :: Monad m => Fix m -> m x
+loop (Fix x) = x >>= loop
 
 --------------------------------------------------------------------------------
 -- Operations
@@ -183,109 +162,7 @@ emptyBot = pureStatelessBot $ const mempty
 hoistBot :: (forall x. m x -> n x) -> Bot m s i o -> Bot n s i o
 hoistBot f (Bot b) = Bot $ fmap (fmap f) b
 
---------------------------------------------------------------------------------
--- Matrix Bot
---------------------------------------------------------------------------------
-
-type MatrixBot m s = Bot m s (RoomID, Event) [(RoomID, Event)]
-
-liftMatrixIO :: (MonadIO m, MonadError MatrixError m) => MatrixIO x -> m x
-liftMatrixIO m = liftEither =<< liftIO m
-
-readFileMaybe :: String -> IO (Maybe T.Text)
-readFileMaybe path = fmap Just (T.readFile path)
-  `catch` \e -> if isDoesNotExistError e then pure Nothing else throwIO e
-
-matrix :: (MonadError MatrixError m, MonadIO m) => ClientSession -> FilePath -> m (Server m [(RoomID, Event)] [(RoomID, Event)])
-matrix session cache = do
-  -- Setup cache
-  liftIO $ createDirectoryIfMissing True cache
-  since <- liftIO $ readFileMaybe $ cache <> "/since_file"
-
-  -- Log in
-  userId   <- liftMatrixIO $ getTokenOwner session
-  filterId <- liftMatrixIO $ createFilter session userId messageFilter
-
-  -- Start looping
-  go filterId since
-
+fixBot :: Functor m => Bot m s i o -> s -> Behavior m i o
+fixBot (Bot b) = go
   where
-    go filterId since = do
-      -- Get conversation events
-      syncResult <- liftMatrixIO $ sync session (Just filterId) since (Just Online) (Just 1000)
-
-      -- Unpack sync result
-      let newSince :: T.Text
-          newSince = syncResult ^. _srNextBatch
-
-          roomsMap :: Map.Map T.Text JoinedRoomSync
-          roomsMap = syncResult ^. _srRooms . _Just . _srrJoin . ifolded
-
-          roomEvents :: Map.Map T.Text [RoomEvent]
-          roomEvents = roomsMap <&> view (_jrsTimeline . _tsEvents . _Just)
-
-          events :: [(RoomID, Event)]
-          events = Map.foldMapWithKey
-            (\rid es -> fmap ((RoomID rid, ) . view _reContent) es)
-            roomEvents
-
-      return $ Server $ (events,) $ \outputs -> do
-        -- Send the bot's responses
-        gen <- newStdGen
-        let txnIds = TxnID . T.pack . show <$> randoms @Int gen
-        liftIO $ zipWithM_ (uncurry $ sendMessage session) outputs txnIds
-
-        -- Update since file
-        liftIO $ writeFile (cache <> "/since_file") (T.unpack newSince)
-
-        -- Do it again
-        go filterId (Just newSince)
-
-simplifyMatrixBot :: Monad m => MatrixBot m s -> TextBot m s
-simplifyMatrixBot (Bot bot) = Bot $ \i s -> do
-  BotAction {..} <- bot (RoomID mempty, mkMsg i) s
-  pure $ BotAction (fmap (viewBody . snd) responses) nextState
-
-liftSimpleBot :: Functor m => TextBot m s -> MatrixBot m s
-liftSimpleBot (Bot bot) = Bot
-  $ \(rid, i) s -> fmap (fmap (fmap ((rid, ) . mkMsg))) $ bot (viewBody i) s
-
-viewBody :: Event -> T.Text
-viewBody = view (_EventRoomMessage . _RoomMessageText . _mtBody)
-
-mkMsg :: T.Text -> Event
-mkMsg msg = EventRoomMessage $ RoomMessageText $ MessageText msg
-                                                             TextType
-                                                             Nothing
-                                                             Nothing
-
---------------------------------------------------------------------------------
--- Text Bot
---------------------------------------------------------------------------------
-
--- | A 'SimpleBot' maps from 'Text' to '[Text]'. Lifting into a
--- 'SimpleBot' is useful for locally debugging another bot.
-type TextBot m s = Bot m s T.Text [T.Text]
-
-repl :: IO (Server IO [T.Text] T.Text)
-repl = do
-  -- Read the user's input
-  putStr "<<< "
-  hFlush stdout
-  (T.pack -> input) <- getLine
-
-  return $ Server $ (input,) $ \outputs -> do
-    -- Print the bot's responses
-    traverse_ (putStrLn . T.unpack . (">>> " <>)) outputs
-
-    -- Do it again
-    repl
-
-annihilate :: Monad m => Behavior m i o -> Server m o i -> Fix m
-annihilate (Behavior a) (Server (i, b)) = Fix $ do
-  BotAction { nextState = a', responses = o } <- a i
-  b' <- b o
-  return $ annihilate a' b'
-
-loop :: Monad m => Fix m -> m x
-loop (Fix x) = x >>= loop
+  go s = Behavior $ \i -> first go <$> b i s
