+++ src/CofreeBot/Server.hs
@@ -1,0 +1,10 @@
+-- |
+module CofreeBot.Server
+  ( module Matrix
+  , module Repl
+  , module Type
+  ) where
+
+import CofreeBot.Server.Matrix as Matrix
+import CofreeBot.Server.Repl as Repl
+import CofreeBot.Server.Type as Type
